Quick API Refactor
I misread the take-home exercise prompt, and had previously been
returning a random suggestion from the /v1/suggestions endpoint, instead
of a list of all of the records.

So in this commit, I changed the mirage server to reflect this, and
changed the request flow. On page load, all suggestions are fetched from
the API, and cached in component state. Then every time a user spins, they'll
see a random suggestion from the suggestions stored in the component state.